BAU: Create ECR repository for each image
After working on HOTT-3513, I have come to the realisation that we do
need a repository for each separate image. This is why I am getting the
error 'EOF' from docker push commands.

Thus, I have changed the ECR module to create a repository for each of
the applications; frontend, backend, duty-calculator, and admin.

Whilst here, I've also changed resource names to 'this'; they are in
a module and there is only one of each resource type, so do not need
special names, as they'll be referenced by the module call in each
environment they are used in - this also conforms to best practices.