BAU: Allow HTTP traffic through to ECS from ALB
[ci skip] because I don't want this to apply over the other open PRs

We want HTTP traffic to be reach the applications from the load
balancer, because the applications health checks are done by the load
balancer - and can only be done over a non-SSL connection. This is the
simplest way of digging ourselves out of that hole.

I've also removed variables that were only being used once or twice.