BAU: Create all remaining Target Groups
[ci skip]

We need the rest of the target groups at this stage so we can deploy
the remaining applications easily.

I've moved the list of services out into a local with a map for each
service that contains its target group name, the priority of the rule,
and the paths for the rule. This makes editing the rules easier in the
future if we need to add more paths.