HOTT-3733: Create MySQL RDS Instance for Signon
[ci skip] because the changes are quite big.

This PR makes significant change to the RDS module, in that it takes
the connection string secret generation away from having to be created
downstream. It does this by using a data source to fetch the secret
created by RDS and extracting the password and parsing the inputs to
build a connection string.

I have also changed the name of the subnet group as it would otherwise
clash with the new MySQL instance.