FPO-133: Adds priority to the OTT services
Terraform/aws increments the priority based on the insertion order of new rules.
This was leading to the frontend (e.g. more generic) rule matching ahead of the
more specific hub rules which is not what we want.

I've added 10 to the existing priorities to avoid collisions when setting these
new priorities